+++ app/src/main/java/com/example/house_analysis/network/model/request/LoungeFloorModel.kt
@@ -3,5 +3,5 @@ package com.example.house_analysis.network.model.request
 data class LoungeFloorModel(
     val floorNumber: Int,
     val loungeNumber: Int,
-    val subtaskIds: List<Int>
+    val subtaskIds: List<Long>
 )
